+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Changelog and tags
 - Logo
+- Memoization of class methods
 
 ## [0.2.0] - 2020-10-28
 ### Added